Add files via upload
The ACMS aircraft health monitoring tool consists of a multitude of sensor type inputs and data processing algorithms to determine the time of maintenance and predict the probability of failure before flight. The program compares the real-time numerical input of the sensors to the healthy baseline values [4]. The ACMS system utilizes the numerical stochastic uncertainty logic to determine the probability of failure [1].
The previous iteration of this project contained a Jupyter-notebook python file, titled “AircraftHealth.ipynb”. This file contained 5 out of the dozens of the real-life aircraft maintenance variables [9]. 
•	Vibration
•	Temperature
•	Pressure
•	Electric current
•	Rotational speed
A set of fuzzy rules were applied to each of the 5 variables. These rules defined what constitutes a low, medium, and high extreme of values. The three fuzzy rules become the input of the control system which calculates the output value. 
	Old Three Rule Code: 
 
	New Two Rule Code: 
 
The original iteration had a set of three rules that corresponded to: low, medium, and high values. However, the new iteration was revised to utilize a two rule system that corresponds a simple: “yes” or “no” maintenance needed value. The room for uncertainty is also significantly smaller in the new two rule control system when compared to the old three rule system [5].